requirements.txt
Flask
scikit-learn
tweepy
getoldtweets3
langdetect
pymongo
google-cloud-language